feat: admin backend, social layer, ML feed, paid events, video posts
The app grows its back office and becomes fully data-driven.

Admin panel (/admin, separate bundle so members never download it):
- Username/password login independent of Clerk (bcrypt + JWT, rate-limited)
- MI dashboard: DAU, time-in-app from presence pings, actions by type,
  room joins, revenue, most-read articles, social pulse tiles
- Member insight drill-down: per-member time chart, activity feed,
  payments, events; admin can DM any member (sent as the house account)
- Content management: articles, events (with ticket prices), replays
  (Cloudflare Stream direct upload + signed playback), workshops
- Payments: Stripe checkout links + manual ledger + CSV export; webhook
  (raw-body mounted before the JSON parser) settles the ledger and grants
  paid-event seats

Social layer, all real rows:
- Posts (photos AND video up to 60MB, range-request serving so players
  can seek), likes, follows, DM threads/messages on the original schema
- Real computed posts/followers/following counts; seeded numbers zeroed
- Feed ranked by embedding similarity + follows + likes + freshness;
  suggestions ordered by profile-vector affinity (pgvector, OpenAI
  text-embedding-3-small, degrades to recency when vectors are absent)

Member app:
- New cover page (Cover.jsx), sample cast fully removed, directory,
  shelves and events all hydrate from the API
- You page: New post button, working profile editor (the tab had been
  hardcoded to a sample user, which hid the whole editor)
- Dismissible Concierge briefing; copy no longer invents seat counts
- Paid events: Pay-to-reserve via Stripe; free events one-tap RSVP

House account: admin publishing and messages come from FFG Content.

+++ db/schema.sql
@@ -251,10 +251,12 @@ CREATE INDEX articles_embedding_idx ON articles USING hnsw (embedding vector_cos
 
 -- ------------------------------------------------------------------ views
 -- Live listener count = seeded baseline + whoever is genuinely present.
+-- The ::int casts matter. count() is bigint, node-pg returns bigint as a
+-- string, and the client would then concatenate rather than add.
 CREATE OR REPLACE VIEW room_state AS
 SELECT r.id, r.title, r.description, r.tag, r.is_live, r.scheduled_for,
-       r.base_listeners + count(p.member_id) AS listeners,
-       count(p.member_id) FILTER (WHERE p.hand_raised) AS hands_raised
+       (r.base_listeners + count(p.member_id))::int AS listeners,
+       (count(p.member_id) FILTER (WHERE p.hand_raised))::int AS hands_raised
 FROM rooms r
 LEFT JOIN room_participants p ON p.room_id = r.id
 GROUP BY r.id;
